requirements.txt
compressai==1.2.6
decode==2024.9.2
diffusers==0.30.3
helpers==0.2.0
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.3
Pillow==10.4.0
pillow_heif==0.18.0
pytorch_msssim==1.0.0
torch==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tensorboard
accelerate
